packages/it-automation-martin/it-automation-martin-1.1.2.tar.gz/it-automation-martin-1.1.2/tests/test_supplier_image_upload.py
import os.path
import pytest
from unittest import mock
from it_automation.supplier_image_upload import post_images


@pytest.mark.parametrize(
    "_input, expected",
    [(201, "Success"), (400, "POST error status=400")]
)
@mock.patch("it_automation.run.requests.post")
def test_post_images(mock_requests_post, _input, expected):
    mock_requests_post.return_value = mock.Mock(**{"status_code": _input})

    test_url = 'test_url'

    file_path = os.path.dirname(os.path.realpath(__file__))
    test_image_directory = os.path.join(file_path, 'images')

    if _input != 201:
        with pytest.raises(Exception, match=expected):
            post_images(test_url, test_image_directory)
    else:
        post_images(test_url, test_image_directory)
    mock_requests_post.assert_called()
